requirement.txt
fastapi
python-jose[cryptography]
pydantic
pydantic[email]
passlib
sqlalchemy
bcrypt
uvicorn
python-multipart
psycopg2
python-dotenv